Abstracted snowflake connection to SnowflakeTask for #34.
